Move repo analyzer to its own package
In preparation for adding more things to the blobstore testkit, this
commit moves the repository analyzer implementation from
`o.e.r.blobstore.testkit` to `o.e.r.blobstore.testkit.analyze`.
