+++ x-pack/plugin/snapshot-repo-test-kit/src/main/java/org/elasticsearch/repositories/blobstore/testkit/SnapshotRepositoryTestKit.java
@@ -20,6 +20,8 @@ import org.elasticsearch.core.TimeValue;
 import org.elasticsearch.features.NodeFeature;
 import org.elasticsearch.plugins.ActionPlugin;
 import org.elasticsearch.plugins.Plugin;
+import org.elasticsearch.repositories.blobstore.testkit.analyze.RepositoryAnalyzeAction;
+import org.elasticsearch.repositories.blobstore.testkit.analyze.RestRepositoryAnalyzeAction;
 import org.elasticsearch.rest.RestController;
 import org.elasticsearch.rest.RestHandler;
 import org.elasticsearch.xcontent.XContentBuilder;
@@ -51,7 +53,8 @@ public class SnapshotRepositoryTestKit extends Plugin implements ActionPlugin {
         return List.of(new RestRepositoryAnalyzeAction());
     }
 
-    static void humanReadableNanos(XContentBuilder builder, String rawFieldName, String readableFieldName, long nanos) throws IOException {
+    public static void humanReadableNanos(XContentBuilder builder, String rawFieldName, String readableFieldName, long nanos)
+        throws IOException {
         assert rawFieldName.equals(readableFieldName) == false : rawFieldName + " vs " + readableFieldName;
 
         if (builder.humanReadable()) {
